Silence warnings from clang.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@36106 72102866-910b-0410-8b05-ffd578937521

+++ Source/gsc/GSStreamGState.m
@@ -62,13 +62,13 @@
     *linejoin = clinejoin;
 }
 
-- (void) DPScurrentlinewidth: (float*)width
+- (void) DPScurrentlinewidth: (CGFloat*)width
 {
   if (width)
     *width = clinewidth;
 }
 
-- (void) DPScurrentmiterlimit: (float*)limit
+- (void) DPScurrentmiterlimit: (CGFloat*)limit
 {
   if (limit)
     *limit = cmiterlimit;
@@ -90,12 +90,12 @@
   clinejoin = linejoin;
 }
 
-- (void) DPSsetlinewidth: (float)width
+- (void) DPSsetlinewidth: (CGFloat)width
 {
   clinewidth = width;
 }
 
-- (void) DPSsetmiterlimit: (float)limit
+- (void) DPSsetmiterlimit: (CGFloat)limit
 {
   cmiterlimit = limit;
 }
@@ -106,5 +106,3 @@
 }
 
 @end
-
-
